Change Dockerfile to use manual server-start.sh and fix ACL permissions for alice-producer
Hving unexplained issues w/ configurations not picking up when runnning
via the CMD entrypoint.  Works when running the same script on a shell session.
Troubleshooting is too low priority right now; willl come back to this

Added All permissions for Alice producer; should only need a few, not worth checking
ones for sure for the sake of this POC